refactor: LogHelper
