+++ src/app/modules/LogHelper.ts
@@ -1,86 +1,50 @@
-import winston, { Logger } from 'winston'
-import Bottleneck from 'bottleneck'
-import appRoot from 'app-root-path'
-
-const { printf, combine, timestamp, colorize } = winston.format
-
-const printLog = printf(({ level, message, timestamp }) => {
-  return `${timestamp} ${level}: ${message}`
-})
-
-const logColors = {
-  error: 'red',
-  warn: 'yellow',
-  info: 'green',
-  http: 'blue',
-  debug: 'gray'
-}
-winston.addColors(logColors)
-
-const logOptions = {
-  outFile: {
-    level: 'info',
-    filename: `${appRoot.path}/logs/out.log`,
-    handleExceptions: true,
-    json: false,
-    maxsize: 5242880,
-    maxFiles: 10,
-    format: combine(timestamp(), printLog)
-  },
-  errorFile: {
-    level: 'error',
-    filename: `${appRoot.path}/logs/error.log`,
-    handleExceptions: true,
-    json: false,
-    maxsize: 5242880,
-    maxFiles: 3,
-    format: combine(timestamp(), printLog)
-  },
-  console: {
-    level: 'debug',
-    handleExceptions: true,
-    json: false,
-    colorize: true,
-    format: combine(colorize(), timestamp(), printLog)
-  }
-}
+import winstonConfig from '@@/configs/winston'
+import { Logger } from 'winston'
 
 type levelOptions = 'error' | 'warn' | 'info' | 'http' | 'debug'
 
 export default class LogHelper {
-  /**
-   * Rate limiter for logger
-   */
-  private static rateLimiter = new Bottleneck({
-    maxConcurrent: 1,
-    minTime: 5
-  })
+  private static _instance: LogHelper
 
   /**
-   * Winston logger
+   * winston logger
    */
-  private static logger = winston.createLogger({
-    transports: [
-      new winston.transports.File(logOptions.outFile),
-      new winston.transports.File(logOptions.errorFile),
-      new winston.transports.Console(logOptions.console)
-    ],
-    exitOnError: false
-  })
+  private logger: Logger
 
-  public static getLogger(): Logger {
-    return this.logger
+  /**
+   * Get singleton instance
+   *
+   * @constructor
+   */
+  public static get Instance(): LogHelper {
+    if (this._instance === undefined) {
+      this._instance = new this()
+    }
+
+    return this._instance
   }
 
   /**
-   * Log with rate limiting.
-   *
-   * @param level
-   * @param message
+   * Initialize logger
    */
-  public static log(level: levelOptions, message: string): void {
-    this.rateLimiter.schedule({}, () => {
-      return Promise.resolve(this.getLogger().log({ level, message }))
+  private constructor() {
+    this.logger = winstonConfig.winston.createLogger({
+      transports: [
+        new winstonConfig.winston.transports.File(
+          winstonConfig.logOptions.outFile
+        ),
+        new winstonConfig.winston.transports.File(
+          winstonConfig.logOptions.errorFile
+        ),
+        new winstonConfig.winston.transports.Console(
+          winstonConfig.logOptions.console
+        )
+      ],
+      exitOnError: false
     })
+  }
+
+  public log(level: levelOptions, message: string): Logger {
+    return this.logger.log({ level, message })
   }
 }
